Merge branch 'master' into 'athena_22.0.44_qw'

+++ Simulation/SimulationJobOptions/share/specialConfig/preInclude.Monopole.py
@@ -5,33 +5,34 @@
 def load_files_for_monopole_scenario(MASS, GCHARGE):
     import os, shutil, sys
 
-    import ExtraParticles.PDGHelpers
+    from G4AtlasApps.SimFlags import simFlags
+    from ExtraParticles.PDGHelpers import getPDGTABLE
+    if getPDGTABLE(simFlags.ExtraParticlesPDGTABLE.get_Value()):
+        ALINE1="M 4110000                         {intmass}.E+03       +0.0E+00 -0.0E+00 Monopole         0".format(intmass=int(MASS))
+        ALINE2="W 4110000                          0.E+00         +0.0E+00 -0.0E+00 Monopole         0"
+        BLINE1="4110000 {intmass}.00 0.0 {gcharge} # Monopole".format(intmass=int(MASS), gcharge=GCHARGE)
+        BLINE2="-4110000 {intmass}.00 0.0 -{gcharge} # MonopoleBar".format(intmass=int(MASS), gcharge=GCHARGE)
 
-    ALINE1="M 4110000                         {intmass}.E+03       +0.0E+00 -0.0E+00 Monopole         0".format(intmass=int(MASS))
-    ALINE2="W 4110000                          0.E+00         +0.0E+00 -0.0E+00 Monopole         0"
-    BLINE1="4110000 {intmass}.00 0.0 {gcharge} # Monopole".format(intmass=int(MASS), gcharge=GCHARGE)
-    BLINE2="-4110000 {intmass}.00 0.0 -{gcharge} # MonopoleBar".format(intmass=int(MASS), gcharge=GCHARGE)
+        f=open('PDGTABLE.MeV','a')
+        f.writelines(str(ALINE1))
+        f.writelines('\n')
+        f.writelines(str(ALINE2))
+        f.writelines('\n')
+        f.close()
+        partmod = os.path.isfile('particles.txt')
+        if partmod is True:
+            os.remove('particles.txt')
+        f=open('particles.txt','w')
+        f.writelines(str(BLINE1))
+        f.writelines('\n')
+        f.writelines(str(BLINE2))
+        f.writelines('\n')
+        f.close()
 
-    f=open('PDGTABLE.MeV','a')
-    f.writelines(str(ALINE1))
-    f.writelines('\n')
-    f.writelines(str(ALINE2))
-    f.writelines('\n')
-    f.close()
-    partmod = os.path.isfile('particles.txt')
-    if partmod is True:
-        os.remove('particles.txt')
-    f=open('particles.txt','w')
-    f.writelines(str(BLINE1))
-    f.writelines('\n')
-    f.writelines(str(BLINE2))
-    f.writelines('\n')
-    f.close()
-
-    del ALINE1
-    del ALINE2
-    del BLINE1
-    del BLINE2
+        del ALINE1
+        del ALINE2
+        del BLINE1
+        del BLINE2
 
 doG4SimConfig = True
 from AthenaCommon.AthenaCommonFlags import athenaCommonFlags
